[Script] convenience script to add token and query view functions
