+++ scripts/stablex/invokeViewFunction.js
@@ -1,0 +1,6 @@
+const StablecoinConverter = artifacts.require("StablecoinConverter")
+const { invokeViewFunction } = require("../script_utilities.js")
+
+module.exports = async (callback) => {
+  await invokeViewFunction(StablecoinConverter, callback)
+}
